Workaround for build script permission issues
Transfering between windows and unix systems can get the permissions reset for these executables. The sh command just runs the script, whether it is executable or not. The problem can still occur for the top-level make_*.sh scripts, but those are meant for running on unix, so the user can probably fix those themself.